Clean up code using IntelliJ code inspection

+++ src/main/java/client/command/commands/gm6/ShutdownCommand.java
@@ -51,10 +51,10 @@ public class ShutdownCommand extends Command {
         }
 
         if (time > 1) {
-            int seconds = (int) (time / 1000) % 60;
-            int minutes = (int) ((time / (1000 * 60)) % 60);
-            int hours = (int) ((time / (1000 * 60 * 60)) % 24);
-            int days = (int) ((time / (1000 * 60 * 60 * 24)));
+            int seconds = (time / 1000) % 60;
+            int minutes = (time / (1000 * 60)) % 60;
+            int hours = (time / (1000 * 60 * 60)) % 24;
+            int days = (time / (1000 * 60 * 60 * 24));
 
             String strTime = "";
             if (days > 0) strTime += days + " days, ";
